backend/requirements.txt
# Django e dependências principais
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-allauth==0.57.0
python-decouple==3.8

# Autenticação Google
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Processamento de imagens
Pillow==10.1.0

# Banco de dados
psycopg2-binary==2.9.9

# Cache e tarefas assíncronas
celery==5.3.4
redis==5.0.1
django-redis==5.4.0

# Segurança e validação
django-ratelimit==4.1.0
django-extensions==3.2.3

# Monitoramento e logging
sentry-sdk[django]==1.38.0

# Utilitários
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7

# Desenvolvimento (opcional)
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0
